Converted some integer values to decimal

+++ Models/NewbornModel.cs
@@ -9,26 +9,26 @@ namespace NeoHearts_API.Models
         public string? FirstName { get; set; }
         [StringLength(50)]
         public string? LastName { get; set; }
-        public int Age { get; set; }
+        public decimal Age { get; set; }
         [Required, RegularExpression("^(Male|Female|Other)$", ErrorMessage = "Sex must be Male, Female, or Other.")]
         public string? Sex { get; set; }
         [Required]
         public DateOnly DOB { get; set; }
-        public int? Gestational_Age { get; set; }
+        public decimal? Gestational_Age { get; set; }
         //[Required, RegularExpression("^(SVD|VAD|LSCS|Other)$", ErrorMessage = "Mode of Delivery must be SVD, VAD, LSCS, or Other.")]
         public string? Mode_of_Delivery { get; set; }
         public float? Birth_Weight { get; set; }
-        public int? Apgar_Scores_1min { get; set; }
-        public int? Apgar_Scores_5min { get; set; }
+        public decimal? Apgar_Scores_1min { get; set; }
+        public decimal? Apgar_Scores_5min { get; set; }
         public string? Resuscitation { get; set; }
-        public int? Maternal_age { get; set; }
-        public int? Parity { get; set; }
+        public decimal? Maternal_age { get; set; }
+        public decimal? Parity { get; set; }
         public string? Prenatal_ultrasound { get; set; }
         public string? CHD_CCHD { get; set; }
         public string? Extracardiac_Diagnosis { get; set; }
         public string? Sepsis { get; set; }
-        public int HR { get; set; }
-        public int RR { get; set; }
+        public decimal HR { get; set; }
+        public decimal RR { get; set; }
         public float? T { get; set; }
         public string? Cyanosis { get; set; }
         public string? Increased_work_of_breathing { get; set; }
@@ -54,8 +54,8 @@ namespace NeoHearts_API.Models
         public string? Sucking_reflex { get; set; }
         public string? Tone { get; set; }
         public string? Activity { get; set; }
-        public int Right_upper_arm { get; set; }
-        public int Right_leg { get; set; }
+        public decimal Right_upper_arm { get; set; }
+        public decimal Right_leg { get; set; }
         public string? CCHD_first { get; set; }
         public string? Congenital_heart_disease { get; set; }
         public string? Open_heart_surgery { get; set; }
